move package of document validators which closes #769
