+++ redpen-core/src/main/java/cc/redpen/validator/document/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * DocumentValidators are provided.
+ */
+package cc.redpen.validator.document;
