+++ redpen-core/src/main/resources/cc/redpen/validator/section/error-messages.properties
@@ -2,10 +2,6 @@ SectionLengthValidator=The number of characters in the section ({0}) exceeds the
 ParagraphNumberValidator=The number of paragraphs exceeds the maximum of {0}.
 ParagraphStartWithValidator=Found invalid beginning of paragraph "{0}".
 DuplicatedSectionValidator=Found duplicated section (starts from line {0}).
-UnexpandedAcronymValidator.UnexpandedAcronym=The expanded form of the acronym {0} is not present in the document.
-FrequentSentenceStartValidator.SentenceStartTooFrequent={0}% of sentences start with "{1}".
-WordFrequencyValidator.WordUsedTooFrequently=The word "{0}" comprises {1}% of all words. Expected around {2}%.
 SectionLevelValidator=The section "{0}" is too deep.
 VoidSectionValidator=The section is void.
 EmptySectionValidator=The section "{0}" is empty.
-GappedSectionValidator=The section "{0}" (level {1}) is misplaced. It should have level of {2}.
